More bullet-proof way fo detecting if ipv6 is supported or not when using the native transport.
Motivation:

We should try to bind to an ipv6 only socket before we enable ipv6 support in the native transport as it may not work due setup of the platform.

Modifications:

Try to bind to the ipv6 only address and only use IPV6 later on if this works.

Result:

Fixes [#7021].